Add SUBSHELL_START and SUBSHELL_END tokens.
In order to support nested subshells, the contents of a subshell can be
broken into multiple tokens. This made it impossible to distinguish between
adjacent parts of the same subshell, and adjacent parts of disconnected
subshells.

This commit introduces the SUBSHELL_START and SUBSHELL_END tokens to make it
explicit where one subshell ends and another begins.